+++ CHANGELOG.md
@@ -2,6 +2,56 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [1.0.8](https://github.com/apider-coding/isn-poller/compare/v1.0.7...v1.0.8) (2025-10-01)
+
+
+### Bug Fixes
+
+* cleanup ([7f54cc8](https://github.com/apider-coding/isn-poller/commit/7f54cc82e8a6d384c5b29a0b584b815e39de2da8))
+
+### [1.0.7](https://github.com/apider-coding/isn-poller/compare/v1.0.6...v1.0.7) (2025-10-01)
+
+
+### Bug Fixes
+
+* Using the 10.7cm Flux as a Proxy for Sunspot Number ([16e3c5a](https://github.com/apider-coding/isn-poller/commit/16e3c5a17d84ef22e1a19507827f8868d14f9255))
+
+### [1.0.6](https://github.com/apider-coding/isn-poller/compare/v1.0.5...v1.0.6) (2025-10-01)
+
+
+### Bug Fixes
+
+* Convert metric values to strings in Blynk URL construction ([a61c8cc](https://github.com/apider-coding/isn-poller/commit/a61c8cc14f515e438566f09ddf465e859ea19aa4))
+
+### [1.0.5](https://github.com/apider-coding/isn-poller/compare/v1.0.4...v1.0.5) (2025-10-01)
+
+
+### Bug Fixes
+
+* flux to integer ([fc82c4e](https://github.com/apider-coding/isn-poller/commit/fc82c4e7b4aba75f31c9323602810bd9e4f01843))
+* turn off dependabot ([a7069b8](https://github.com/apider-coding/isn-poller/commit/a7069b83c7925bf4a5562055e212a474aa97c37a))
+
+### [1.0.4](https://github.com/apider-coding/isn-poller/compare/v1.0.3...v1.0.4) (2025-10-01)
+
+
+### Bug Fixes
+
+* rename var ([639f593](https://github.com/apider-coding/isn-poller/commit/639f5937b92d5de0c231a5634d928ee7c8e19ff6))
+
+### [1.0.3](https://github.com/apider-coding/isn-poller/compare/v1.0.2...v1.0.3) (2025-10-01)
+
+
+### Bug Fixes
+
+* remove json ([f7a8e0b](https://github.com/apider-coding/isn-poller/commit/f7a8e0b5a5c32afd3cd248d84dc49115c5a04e72))
+
+### [1.0.2](https://github.com/apider-coding/isn-poller/compare/v1.0.1...v1.0.2) (2025-10-01)
+
+
+### Bug Fixes
+
+* update data getting ([4f00049](https://github.com/apider-coding/isn-poller/commit/4f00049b9416ca8b0deea83a8c22fade4b3b1f49))
+
 ### [1.0.1](https://github.com/apider-coding/isn-poller/compare/v1.0.0...v1.0.1) (2023-05-30)